[Version] 1.1.0 beta 1 [Improve] Highlight lyrics animation Lyrics auto scroll smoothness Modify quality name to match offical NCM Cookie pool sampling method when getting music resources [Fix] Incorrct usage of "setTextSize" #6 [Remove] Finished todos

+++ common/src/main/java/indi/etern/musichud/client/music/decoder/AudioDecoder.java
@@ -1,7 +1,7 @@
 package indi.etern.musichud.client.music.decoder;
 
 public interface AudioDecoder extends AutoCloseable {
-    byte[] readChunk(int maxSize);
+    byte[] readChunk(long maxSize);
     int getFormat();
     int getSampleRate();
     void close();
